refactor constructor code. move classes to files
