+++ src/app/lib/constructor/tags/Logic.ts
@@ -1,0 +1,71 @@
+// Copyright 2017 The genesis-front Authors
+// This file is part of the genesis-front library.
+//
+// The genesis-front library is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The genesis-front library is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the genesis-front library. If not, see <http://www.gnu.org/licenses/>.
+
+import { TProtypoElement } from 'genesis/protypo';
+import Tag from './Tag';
+
+const LogicTagNames = {
+    'getvar': 'GetVar',
+    'setvar': 'SetVar',
+    'addtoolbutton': 'AddToolButton',
+    'linkpage': 'LinkPage',
+    'and': 'And',
+    'calculate': 'Calculate',
+    'cmptime': 'CmpTime',
+    'datetime': 'DateTime',
+    'now': 'Now',
+    'or': 'Or',
+    'code': 'Code',
+    'chart': 'Chart',
+    'forlist': 'ForList',
+    'menugroup': 'MenuGroup',
+    'menuitem': 'MenuItem',
+    'qrcode': 'QRcode',
+    'address': 'Address',
+    'appparam': 'AppParam',
+    'data': 'Data',
+    'ecosysparam': 'EcosysParam',
+    'jsontosource': 'JsonToSource',
+    'langres': 'LangRes',
+    'range': 'Range',
+    'sysparam': 'SysParam',
+    'binary': 'Binary',
+    'settitle': 'SetTitle',
+    'inputerr': 'InputErr',
+    'select': 'Select',
+    'inputmap': 'InputMap',
+    'map': 'Map',
+    'include': 'Include'
+};
+
+class Logic extends Tag {
+    constructor(element: TProtypoElement) {
+        super(element);
+        this.tagName = LogicTagNames[element.tag] || element.tag;
+        this.canHaveChildren = false;
+        this.logic = true;
+        this.attr = {
+        };
+        this.editProps = [];
+        for (let attr in element.attr) {
+            if (element.attr.hasOwnProperty(attr)) {
+                this.attr[attr] = attr;
+            }
+        }
+    }
+}
+
+export default Logic;
